fix: 경기 상세 조회와 랭크 전적 필터링 개선

+++ src/main/java/com/toyproject/eron/erapi/dto/GameDetailResponse.java
@@ -4,7 +4,7 @@ import java.util.List;
 import java.util.Map;
 
 public record GameDetailResponse(
-        Integer gameId,
+        Long gameId,
         Integer seasonId,
         Integer matchingMode,
         Integer matchingTeamMode,
